cleanup of build files.
rename candlepin-home to candlepin.home.
removed basedir from <project ...> entry so that it defaults to basedir
of the build.xml file. This was needed because we switched to build.properties
so what used to be set via property ... location= no longer worked.
commented out currently unneeded jars from ivy.xml