Add an isFree column to the entitlement table.
May not be necessary but easier to remove later than it would be to add.
Tracks if the entitlement was granted for free or not, i.e. for virt
guests getting entitlements for free. May help us clean up in the event
the parent loses an entitlement that allowed this one to be granted.