Basic Auth for the consumer create method
The authenticated username is passed to the create method by means of a request
attribute.  We weren't able to find a way to set the principal.

The jdbc lookup hardcoded to the web db, for testing only.  We weren't exactly
sure how this should be configured.

NOTE: If you apply this commit the tests will need to be updated since the
Consumer constructor signature has changed.  I wasn't sure if other methods
would need the request object and the constructor seemed like the most logical
place to put it.  If there is a better place to inject the request feel free to
change it.