Drop the Pools object.
This looks wrong, the bug referenced doesn't seem applicable to our
scenario, and the output looks exactly the same in both XML and JSON as
far as I can tell. Until a reason surfaces again, going to kill this
off, as it's unplesant to start a trend where we need an object for
every collection of other objects we need to return.