change rootLogger to output WARN instead of DEBUG.
Hibernate was configured to use log4j recently, and our rootLogger
was configured at DEBUG causing EVERYTHING to log.

   rootLogger = DEBUG - 2m22.025s

Changing rootLogger to WARN helps a lot:

   rootLogger = WARN - 31.567s

Changing the output format to use %c instead of %C:

   ConversionPattern=...%c... - 30.461s

This is still with org.fedorahosted.candlepin=DEBUG so we
still see all of our debug output if we want.