replace hash with release, use specfile as source.
Instead of using the git hash, we'll simply use the version/release
from the spec file which tito increments during a release for us
automagically. While using the git hash approach is certainly nice
it doesn't help if you're building it outside of a git checkout
say a tarball or rpm.