Test changes for Principal dependency injection.
Difficult to have Guice allow individual tests to override bindings on
the fly. Introduced a singleton that the TestPrincipalProvider will use
to create principals. Tests can also get access to this singleton to set
the Principal they want returned on the fly.

Added overridable Guice config in tests for modules that wish to define
their own Guice config, but do not want to define everything
they need.