json: use iso8601 datetimes
Create our own json provider for jax-rs that's a subclass of the jackson
one. Set up this provider to use iso8601 style datetimes, and provide a
config for pretty printing json.

The configuration could be done with the default provider, but the
provider gives us a good place to do input validation.