Just grab ApplicationException, so RESTeasy can handle its own exceptions
RESTeasy uses a special exception for the HTTP OPTIONS call to provide a
default response if your resources do not implement their own. We want
to make use of this, so we need to let RESTeasy deal with the exception.
So just grab ApplicationException, which wraps all the exceptions that
come from our code.

Note that their are other exceptions we might want to deal with, like
exceptions relating to message reading and writing. We'll have to deal
with those on a case-by-case basis.