Consumers unentitled when quantity of a subscription decreases implemented
Completed implementing the story where when quantity of a subscription
changes and the url /owners/key/subscriptions/ is hit with a 'PUT',
then the pools are refreshed and during the process if the total quantity
consumed from pool exceeds its current quantity, excess entitlements
are automatically revoked.
Commit also includes junit test cases.