PoolManager created. Changes in Subscription dates re-creates entitlement certs
Moved some of the business code/methods from PoolCurator into PoolManager.
This solves the Entitler <-> PoolCurator problem. Also, new test cases
added for the manager; references to pool curator updated in some of the classes.

Changes in date of a Subscription now causes entitlement certificates to be
regenerated and the old ones to be revoked when refreshEntitlementPools
is called. Existing code updated to reflect these changes.