Add server side support for flex expiry.
When present on a product or pool, this attribute should trigger the
post rule which will modify the entitlement objects flex expiry days.
Candlepin then knows to use this value when generating the entitlement
certificate. The certificates end date will be the extended one, while
the order namespace will still contain the earlier subscription end
date.