ruby: Cleanup product creation.
Needed to get rid of the various parameters which were actually meant to
be attributes now. Child products was completely unused. In general we
should be able to create products without having to specify a bunch of
optional args just to get to the one we want, so began adopting a params
hash approach.

Watchout for constraint violations when functional testing, now if you
try to re-create a product with a name that already exists in the db
(but with a different product ID), you will see this happen. (easy to
fix)