Add HATEOAS support.
Registering a custom serializer for entities we consider linkable within
Candlepin. This serializer will examine the context in which the entity
is being serialized. If it is within the serialization of some other
object, we will switch to serialize just the objects 'id' and 'href'. If
in a collection, we will do the same. If in the root context however, we
proceed with normal object serialization.

This normal serialization passthrough involves some brutal Jackson
hacks, hopefully this may be addressed in the future with some patches
upstream.