Emit pool creation events.
This was already happening for straight pool creation events, but not
for during subscription refresh or sub-pool creation.

Involves a huge refactor to help with Guice circular dependency
problems. Entitler and PoolManager were turning into a problem, but with
functionality similar enough to be one class, so they have both merged
into PoolManager.

The PostEntHelper was also causing problems, turned this into a
non-Guice object, as it requires things created at runtime anyhow.