brew prep - add ant build file and trivial support scripts.
Ant will *ONLY* be used to build the candlepin in brew. Buildr's
buildfile is still the master build file of record. The build.xml
will print out a friendly reminder to use buildr if you try to
run it without arguments.

The build.xml does all of the necessary parts to build:

 * candlepin-api.jar
 * candlepin.war
 * generate po files
 * get version and release from the spec file

I didn't bother with rspec, cucumber, or unit tests.