autobind: check all possible pool combinations
For selecting the 'best' pools, we have N pools to choose from, to
satisfy M products. We already know that each of these pools provides at
least one of the requested products, has available entitlements, and is
relevant to the machine (for arch rules, etc), as they have been
filtered prior to arriving at the select best pools call. To pick the
best combination of pools, we:

- Iterate over the pools, figure out which of the M products they each
  provide, and select only 1 of each pool that provides the same set of
  products (right now we pick the pool that expires first).
- Form all possible combinations of pools. For example, if we had the
  pools [A, B, C] our list of combinations would be:
  [A]
  [B]
  [C]
  [A, B]
  [A, C]
  [B, C]
  [A, B, C]
- Remove from the list of possible combinations any combinations which
  provide a product more than once, unless the product allows multi-entitle.
  This is still TODO.
- Iterate over the remaining combinations, and select the one which
  provides the most requested products. If there is a tie, select the
  one which uses the fewest entitlements.