665128: Store product name with pools reliably.
Previously this was acting strange and initially leaving product name
empty, but then setting it as soon as the pool was viewed/used in any
way over the REST API. We now store the product name when the pool is
created from a subscription.

Added checks for changes to the product name during refresh pools and
updates if necessary.