Refactor import products script and add test data.
Import script now expects json formatted with more named keys and less
positional arrays. Settings which are often the same are now defaulted
and do not have to be in the json itself.

Now set TESTDATA=1 when running deploy script to get a selection of test
data populated on your Candlepin server.

Adding a file with a variety of products which should cover the major
use cases Candlepin is designed to handle.