674078: Take a full ISO 8601 timestamp for pool search activeOn
Previously pool search took a YYYYMMDD string for querying active pools.
There were two problems:
 - the comparison was lt and gt, so a pool that was active at the very
   beginning of a day would not count as being active on that date
   (because the Date object's h/m/s were all set to 0)
 - queries from anyone in a timezone shifted away far enough from the
   timezone of the candlepin instance could get results they weren't
   expecting.

Take a full ISO 8601 timestamp as the argument, so a caller of the api
can be precise about the time they are querying.

Note: this breaks the api, since the old YYYYMMDD string will be
interpreted as some year past 9999999 AD