671195: Use pessimistic locking on pools during bind.
Obtain an explicit pessimistic lock on the pool object during
entitlement. Even though we don't actually modify the pool, this will
block subsequent requests to entitle from that pool until the
transaction completes.

Subsequent requests will block waiting for the lock to be release, at
which point they will be allowed to read the pool from the database and
proceed with entitlement.

This lock does not affect operations like list pools, and is only used
during entitlement.