precompile and share our javascript rules across threads
The JsRulesProvider handles reading the rules from disk/database and
compiling them into a rhino global rhino scope (which also contains
standard javascript objects). This scope is marked as sealed, so that
the shared values cannot be altered.

For each thread/http request, we then get a new JsRules object which
is built from a subscope of the global scope. This new subscope contains
the variables and execution context, and is relatively cheap to create.

recompiling the rules when newer ones are found is still todo.