Update unit tests for refresh logic in Javascript.
Responsibilities have changed now and the pool manager no longer
actually checks if the subscription and pool are different, it just
hands this off to the rules and expects to be returned a list of updated
pool objects. This is why some tests are being removed or made to do
less, as they no longer make sense.

Need to add a new suite of tests for the javascript rules themselves
soon.