681287: Exclude pools with a rule when listing for a consumer.
When using listall plus a consumer UUID, we were ignoring the consumer
UUID and instead, just listing all for the owner. This however is not
expected client side where we don't want to see pools that we literally
cannot consume from.

Modified to always run the rules if a consumer UUID is specified in the
request. If listall is used, we'll include warnings from the rules, but
exclude errors.

Old behavior of listing everything for an owner can be accomplished by
specifying an owner ID, or just GET'ing /pools. (the filters will block
anything the authenticated consumer should not see)