organize items as exportXXX methods to be a bit more clear.
I moved the owner export out of execute() into an exportOwner
method. Then I put pools and entitlements in their own methods as well.

Lastly I changed the if statement to throw an exception if
the status code was *NOT* OK, otherwise just do the logic. This removed
the else clauses.