subpackages should require candlepin.
Before I had candlepin require candlepin-webapp, and had each of
the subpackages Provide candlepin-webapp. This seemed like a good
idea at the time, but the problem is when you yum install candlepin
you pull in the first subpackage that provides candlepin-webapp.
This resulted in candlepin-jboss being selected by default.

Only way for this to work required you to install the webapp
of choice i.e. candlepin-tomcat6 or candlepin-jboss plus candlepin.

By making the subpackages Require candlepin, and removing the
Provides: candlepin-webapp, we can now simply install candlepin-XXX
where XXX is a servlet container of choice.

For example, yum install candlepin-tomcat6.