684350: Stop creating new owners during basic authentication.
A side effect of logging in with a user account returned from the user
service, with an owner who does not yet exist on this Candlepin
instance. This was causing problems for some deployments, so now we will
stop creating the owner but still set a detached owner on the principal
and allow the request to proceed.

We do however wish to still create a new owner during consumer
registration if the authenticated principaĺs owner does not yet exist in
our database