Add a new principal and role for trusted external systems
Other applications/systems may wish to communicate with candlepin and
act as themselves, rather than on behalf of a user or consumer. Extend
the oauth code to fall back to this trusted system principal after oauth
authentication has passed, and if there are no headers present for
user/consumer types.

Use this new role for Subscriptions listing, as well.