Allowing the Subject Key Identifier to be injected
This is needed for the hosted entitlement platform.  To work around the
Akamai's certificate size limitation we are going to place all data needed for
authZ in a separate OID.  To make things more complicated Akamai is unable to
read custom OIDs so we are overloading the Subject Key Identifier.  To my
knowledge that is only important when creating a new CA, not for client
certificates.

I have verified that doing this doesn't break client certificate auth with
Akamai.  The default impl that ships with Candlepin will do the same thing as
before.