Remove assumptions that we have subs in our db.
Need to remember that subscriptions and products may not actually live
in our database, so any operations on them need to go through their
respective adapters rather than curators.

The one assumption we can make is that they'll both be coming from our
db if either is. (i.e. it's save to query product's from the curator
inside the default subscription service adapter)