Remove the AccessControlInterceptor.
This is the interceptor that kicks in on all curator calls and db
queries, using the filters on the model classes, to prevent returning
any results you should not be able to see.

While a nice secondary layer of security, we should be catching
everything in the first. (resource level) The complexity involved with
this was too much to maintain particularly during massive refactors like
this, as this interceptor is constantly kicking in when we're well into
system internals doing things that should just work, without having to
hack around it.