Add a trusted user principal implementation.
Old implementation was passing through to regular user auth classes,
doing a user lookup, and then trying to create a user principal. This is
not desired for oauth authentications as the request is likely from
katello, when we're not even managing users locally, and expect to just
trust the request for whatever katello asked for.