Refactor registration to not require a user service lookup normally.
Now we just deal with the username if possible, and attempt to perform
the user lookup only in the case of person consumer registrations, when
we need to check that the user has access to the destination owner. If
the user service does not implement this lookup call (as will be the
case with Katello), handle the exception and assume Katello did this
check for us.