Cancel jobs and pause/unpause scheduler
This allows job cancellation, as well as being able to pause/unpause the job
scheduler. A new job was added, CancelJobJob, that runs every five seconds. If
it detects any jobs in CANCELLED status in cp_job, it will attempt to delete
them from the scheduler. In most cases, this is a no-op, since most shards
won't be aware of the job in their quartz scheduler. A new field on cp_job,
jobGroup, was added as well to facilitate job cancellation.

Additionally, we can pause and unpause the scheduler. When paused, any queued
jobs from users will run at unpausing, but cron style jobs will need to wait
until the next runtime to be executed (this is controlled via the misfire
property on the job).