Update supported products when updating consumer.
Add support for updating a list of installed packages in PUT
/consumers/{uuid}, in addition to the facts list. (which was about all
we updated previously)

This required a change to make the two sets we're interested in updating
be null if the object was instantiated with the empty constructor. This
is so the REST resource can do an update, and know what is being updated
based on what is present in the request.  null collection implies that
information wasn't included in the PUT, whereas if it's populated but
empty, we can know we need to clear the set.

Consumers queried from Hibernate will get an empty set if they have no
facts/installed products.