Add an autoheal attribute for consumers.
This is an instruction to the client to perform an autoheal request,
rather than anything interpreted server side.

The default is true, but it is not initialized as such as far as the
Consumer class is concerned as we need to be able to check if it's null
in an update request.