735087: If quartz is in clustered mode, we shouldn't schedule any jobs.
Doing so will cause a primary key violation on the second
Candlepin node. The primary key for the table is:
trigger name + trigger group.

At first I believed it was that we weren't using unique enough
names. The name collision was a red herring, the real problem
is we don't want to create new jobs when in clustered mode.
We now check to see if quartz is in clustered mode, and if there
are any jobs already scheduled from another node. If not, we
proceed to schedule the jobs as normal. Otherwise, we skip
initialization of the jobs and join the existing cluster.

Added getClassName(String) to Util.
Updated unit tests.