use a list of lists of pools for select_best_pools
before, we'd create a list of available pools, keeping only one of a
pool from each class, where a class is a grouping of pools that all
provide the same products relevant to what was requested. Begin the
process of grouping the pools into classes, and keeping them around, so
we may use them later for stacking.