Owner info lists all the enabled consumer types
Tracking a list of enabled consumer types, but it's not something we use or enforce internally, only there for integrating management applications.
The list of enabled consumer types is calculated by examining all the owners subscriptions, looking for an attribute containing the list of enabled consumer types,
   and adding them to the list.
No entitlements are required, just the existence of the subscription itself.