teach select_best_pools to heal partial stacks
Once select_best_pools has come up with all pool combinations, and we
begin looping over the pool classes to select which pool from a class
we'll use for a combo, we must first examine the partial stacks on the
system. For each partial stack on the system, we build a map of product
id to stack id, and stack id to covered sockets (product sockets
attribute * quantity). This stack id to covered sockets map previously
existed when not considering existing stacks; it's just now extended to
begin at some number > 0 to represent what already exists on the system.

Now we can look at the potential pools we can use for the system. For
each pool, we must look at all products it provides. If any of them map
to an existing partial stack on the system, note the stack id. After
we've looked at all products, we can rule out this pool if it either
provides more than one stack id (which probably indicates human error
when populating product info), or it provides a different stack id for
a product that already is on the system in a partial stack (so we could
not use it to complete the stack).

At this point, we either have determined that this pool doesn't map to
any partial stacks on the system, in which case we proceed as before,
determinine how many entitlements we need to cover all sockets on the
system, or we have determined that it maps to an existing partial
stack, so we begin with a number of covered sockets > 0.