Use host lookup in rules when checking virt_only pools.
Switch the code to use the runtime lookup of host for a consumer instead
of the parent-child consumer relationship, which is getting removed.

Also assume that if virt_only is present, and the source entitlement is
set on the pool, then the guest's host must match the source entitlement
consumer.