Drop virt_limit_spec.rb.
This is no longer valid as the behavior being tested is only if the
config specifies we're *not* in standalone configuration, but our
default says that we are. Rather than flipping config and restarting
tomcat as part of testing, these should be moved to JsPoolRulesTest
where we can control the config progrematically.