Remove parent restricted column in cp_pool.
We shouldn't be adding new columns to the db for JS rule logic when we
have a generic attributes framework for this purpose.

Fixed JS logic to stop assuming virt_only + a source entitlement means
check for a host. Instead added an explicit requires_host pool
attribute, and JS rule to enforce it.