735354: dropdb in deploy script fails on first run
do not exit if dropdb fails. There is no real good way to detect if
dropdb failed because of lack of schema to actual drop, postgresql is
down or being accessed by another user. So we'll just ignore the
exit code and rely on createdb to catch the failures.
If createdb fails because psql is down or being access by another user,
deploy script will fail.