746035: set entitlement start dates to start date of pool
Previously, they were set to "now", which meant subscribing to
pools that started in the future got the wrong entitlement
start date. Which meant that any checks of this "is this
entitlement valid _now_" would be wrong. For this particular
bug, that meant autosubscribe would consider a product entitled
even though it is only entitled for the future.