688707: capture all of the internal exceptions, transform into JSON.
For some unknown reason, when RESTEasy handles an internal exception
it completely bypasses the defined servlet filters which means that
the custom header put in place will not be set.

Another issue is that RESTEasy will not invoke exception mappers
for super classes until *AFTER* it has processed its internal
exceptions. In order to override that functionality, you must have
an exception mapper for *ALL* of the specific internal exceptions.

Once, those are covered a generic RuntimeException mapper can do
the rest of the work for you.

The Response seems to get committed before the VersionFilter
has a chance to add the header during exception mapping. So
a PostProcessInterceptor called VersionPostInterceptor was
added to add the header on successful cases, thus rendering
the VersionFilter obsolete.

VersionFilter is no longer needed REMOVED.