Behavior of unlimited bonus pool when physical pool is exhausted.
In a hosted system, when the physical pool is exhausted, the bonus pool entitlements will be revoked
and the bonus pool quantity will be set to zero. This will block future entitlement consumption from
the bonus pool. The bonus pool quantity will be reset to -1 when entitlements are returned to the
physical pool and normal unlimited behavoir will resume.

Updated the logic for bonus pool quantity manipulation and entitlement revocation.