754843: Fix legacy virt bonus pools missing pool_derived.
When pool_derived was added there was no effort made to upgrade existing
virt bonus pools in hosted deployments. These pools never get
updated/fixed if anything changes on the parent subscription.

To fix, check for what should be a safe state that also indicated this
is a legacy derived pool (virt_only true, virt_limit 0, pool_derived not
set), and if detected we add pool_derived and make sure the pool will
get saved to the db. After this the normal refresh kicks in and the pool
will get modifications to it's parent subscription.

This is probably a temporary fix. In theory we can deploy it, refresh
all owners, and then remove the code for good.