769644: Disable system wide checking of last manifest import date.
This is blocking imports in multi-tenant deployments if one org imports
a manifest, and then another tries with a manifest that is even slightly
older.

I believe this code was intended to prevent the version of the rules and
product definitions from jumping around depending on who imports last.
Rules import is currently disabled and being discussed. As noted in bug
with this patch, we accept that product definitions *could*
theoretically jump around, if they were changing upstream and people
were importing newer/older versions.