Clear up some database related memory leaks in our tests
Each test method (not class!) in a class that inherits from
DatabaseTestFixture instantiates its own full database stack, from
warp-persist through hibernate down to a complete in-memory hsqldb.

This understandably takes time and memory, and there are unfortunately
some leaks going on, keeping a large portion of the objects we create
still alive, after a test completes. A recent change to the schema means
we're now hitting the 1gig memory limit we run our test jvm with.

This commit attempts to clean up some of the leaks, but does not do
enough. There are still two outstanding leaks, as detected by eclipse's
memory analysis tools:
 - org.hibernate.impl.SessionFactoryImpl, referenced in ThreadLocal
 - com.wideplay.warp.persist.PersistenceFilter; I'm not sure what is
   holding all these refs.

Going forward, unless we can eliminate these leaks, we'll have to bump
the max memory on the test jvm, and continue to migrate any test methods
out of classes that use the database, if the method itself does not
require it.