750307: Prevent duplicate pools via unique constraints
Rather than attempt to ensure only a single refreshPools job runs at a
time, which turned out to be as prone to race conditions as the
refreshPools job itself (IMO we would have had to have done a full table
lock on the JobInfo table for any action on JobInfo or quartz jobs),
allow concurrent refreshPools jobs to run, but implement the facilities
we need for optimistic locking on pools, to ensure data integrity.

In particular, this means we have to be able to detect when another
refreshPools job has altered a Pool before us, either by deleting it
(logic already existed), creating it, or modifying it. In any of these
cases, we're altered of another refreshPools job changing the pool
through a javax.persistence exception.

To detect concurrent pool modification, add a @Version column to Pool.
Hibernate takes care of incrementing this for us. On a commit, if the
version is greater than expected (in other words, someone else has made
a commit on the same row), the commit will fail.

To detect concurrent pool creation, use a composite unique constraint on
subscriptionId and a new column, subscriptionSubKey. Since the
javascript pool creation logic can create more than one pool from a
single subscription, we need the subKey to differentiate. The javascript
logic sets this key as well, to distinquish between a vanilla pool, and
a pool to be used by virt guests.

When a refreshPools job fails execution due to a persistence exception,
we tell quartz to rerun it. Rerunning jobs not only ensures that we have
consistant data, but that it is also in sync with the latest
subscription data.

Concerns: There is no way to stop a job from rerunning, and all jobs
will be rescheduled if they fail due to a persistence exception. IMO
this should be ok. In the worst case the job keeps failing, which
would be the same as if it wasn't rerun (except for some computing
resources lost to running the job). In the best case, the job is rerun
and completes successfully.

Optimistic locking on pools could cause some issues with any other code
paths that modify pools. I've only been able to identify the bind/unbind
logic as places where this happens, and they already grab an exclusive
lock on the row before altering it, so they should be fine.