786730: multi-entitlement guests were not migrating properly, and logging fix
There are two issues that this commit fixes:

First, exceptions caught by CandlepinExceptionMapper were not being logged. A
logging statement has been added to correct this.

Second, there was a loop body in revokeGuestEntitlementsMatchingHost that was
modifying the iterator in certain circumstances, causing a
ConcurrentModificationException. This would pop up during guest entitlement
revocation checks that involve multiple entitlements. Instead, we create a list
of entitlements to be revoked and remove them at the end, so we don't tamper
with the iterator while it's in use.