790751: validate older import check against export creation date
Previously, we were saving the manifest creation date as the date when the
import occured into the local CP. This caused issues if a user performed an
import with an older manifest, and then a followed up with an import from a
newer manifest. For example, if manifest A was created in August, and manifest
B was created in November, if a user attempted to import A and then B in
December, this would incorrectly fail.