788940: allow products to have the same name
In Katello, users may want to create the products with the same name in two
orgs. This usually happens during multi-tenant environments.

This removes the constraint, which allows products with the same name to be
committed. Additionally, the create() in ProductResource was previously not
failing when products with the same ID were created, and would simply return
the existing product. Duplicate calls to create() will now fail; the spec tests
have been updated to not attempt to create duplicate products.

NOTE: this removes the "unique" db constraint on the name field in cp_product.