798227: revoke guest certs from previous host on migration
Since we can't ensure the order in which two hosts will report guest
status updates during a migration, we need to consider the case where
the old host has already removed a guest before the new host reports
having a new guest.

In this case, the guestId entry will exist for a consumer, but there
will be not matching entry in a host for it (similar to if the guest had
reported its existance before any host had reported it). We can then
just remove any entitlements that this consumer has, are virt-only, and
are from a pool created by a host that isn't the new host.