Return a 404 instead of 500 on attempts to undo nonexisting imports
Previously, if a user attempted to delete an import that didn't exist, a 500
error would result. Instead, throw a 404 if the manifest is not found.

Additionally, log a message at the end of an import, so we can start to get a
better idea on how long imports take (related to bz #807468).