804071: CRL Revocation Task was not running in hosted.
The only way to refresh scheduled jobs was to delete all of the entries in the quartz tables and restart a node in the cluster.

Now a job can be refreshed by changing the cron string for scheduling in the configuration file. This will cause the initial job to be deleted and a new instance added. Also available is the ability to remove the entries in the quartz tables for a single job type via SQL and have the job reinstated. This will correct other meta-data issues that are not schedule related. These changes still require a single node restart.

This change does not include a way to remove a job by altering the configuration. If the configuration is altered to omit that task, then it will not be replaced after the task is removed from the database via SQL.

The final note is that the last machine that is instated is the configuration that wins. It will be up to the administrator to ensure that they match. This is no different than past behavior.