811581: cannot unregister hypervisor when >1 guests are consuming bonus subs
Previously, if a hypervisor was providing bonus subs to more than one guest,
and then the hypervisor was unregistered, candlepin would throw a 500 error.
This was due to something inside of a for loop removing objects from the
iterator before the loop reached its end.

The for loop has been rewritten to prevent this from happening. Hypervisors
with >1 guest can unregister without issue.