Fix quartz locks db error.
The update script for adding these rows was not correct and was only
inserting the last row in the database. Need to do a separate insert
statement for each row in liquibase XML.

This change is only used when creating new databases, and because the
package hasn't yet been tagged, this change should apply ok.