802263: autodetection of deleted hypervisor cleanup
Previously, if a user deleted an auto-detected hypervisor, the hypervisor would
then be re-detected and re-created, which caused issues since the hypervisor
was also already marked as deleted.

Instead, block deleted hypervisors from being re-created. Additionally, expose
an API for removing the deletion record of the hypervisor, in case it was
deleted by mistake. This is only available for the admin user, since it is not
something that should be done on a regular basis.