Load default consumer types via database instead of init URL.
We were not properly set up for adding new consumer types to a deployed
candlepin being upgraded, they would only appear if it was a clean
install.

Moved consumer type creation to a liquibase script and added to the
create changelog.

The admin user still needs to be created in our init URL, changed the
logic here to create if using default user service adapter, and no
accounts exist. Solution here uses a new method on our user curator (not
adapters we shouldn't force all implementations to add this) to load
just the user count (not all the objects), preventing any DoS issues for
those using our adapter.