Let running jobs finish before shutting down pinsetter
Previously, if pinsetter received a shutdown signal, it would terminate in the
middle of running jobs. IT requested that instead, we let running jobs finish.

The easiest way to test this patch is to add a Thread.sleep() to the refresh
pools job, fire off some jobs via cpc, and then shut down the container. Note
that if you shut down tomcat6 via the service script on fedora, it may simply
pause the connectors and terminate, which will not fire the shutdown hooks. I
am not sure why this happens, but to get around it, simply send a SIGTERM
signal to the tomcat process.