findbugs: Method may fail to clean up stream or resource
This method may fail to clean up (close, dispose of) a stream, database object,
or other resource requiring an explicit cleanup operation.

In general, if a method opens a stream or other resource, the method should use
a try/finally block to ensure that the stream or resource is cleaned up before
the method returns.