821532: Fix create-changelog error from previous fix.
The add column changeset went into create as well as update, but the
column was already included in create causing errors double creating the
column.

Even worse, we can't just remove the changeset from create as it'll just
try to reapply on the first upgrade.

Instead add precondition to only run if the column doesn't already
exist, and if it does, mark as if we've run it.